+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * New cop `EmptyLines` tracks consecutive blank lines
 * New cop `WordArray` tracks arrays of words.
 * [#108](https://github.com/bbatsov/rubocop/issues/108) New cop `SpaceInsideHashLiteralBraces` checks for spaces inside hash literal braces - style is configurable
+* New cop `LineContinuation` tracks uses of the line continuation character (`\`)
 
 ### Bugs fixed